GPU: Add support to dump/load TPC ZS pages in standalone benchmark
